requirements.txt
asgiref==3.3.4
Django==3.2.4
django-ckeditor==6.1.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
Unidecode==1.2.0
